+++ app/src/main/java/com/sdex/activityrunner/app/ActivitiesListAdapter.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.FragmentActivity
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestManager
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.bumptech.glide.request.RequestOptions
@@ -16,19 +17,18 @@ import com.sdex.activityrunner.R
 import com.sdex.activityrunner.databinding.ItemActivityBinding
 import com.sdex.activityrunner.db.cache.ApplicationModel
 import com.sdex.activityrunner.extensions.resolveColorAttr
-import com.sdex.activityrunner.glide.GlideApp
 
 class ActivitiesListAdapter(
     activity: FragmentActivity,
-    private val application: ApplicationModel,
 ) : ListAdapter<ActivityModel, ActivitiesListAdapter.ViewHolder>(DIFF_CALLBACK) {
 
-    private val glide = GlideApp.with(activity)
+    private val glide = Glide.with(activity)
     @ColorInt
     private val exportedColor = activity.resolveColorAttr(android.R.attr.textColorPrimary)
     @ColorInt
     private val notExportedColor = ContextCompat.getColor(activity, R.color.red)
 
+    var application: ApplicationModel? = null
     var itemClickListener: ItemClickListener? = null
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -59,7 +59,7 @@ class ActivitiesListAdapter(
     ) : RecyclerView.ViewHolder(binding.root) {
 
         fun bind(
-            application: ApplicationModel,
+            application: ApplicationModel?,
             item: ActivityModel,
             glide: RequestManager,
             @ColorInt exportedColor: Int,
@@ -71,7 +71,7 @@ class ActivitiesListAdapter(
             binding.packageName.text = item.componentName.shortClassName
             binding.label.text = item.label
             binding.label.isVisible = !item.label.isNullOrBlank() &&
-                item.label != application.name && item.label != item.name
+                item.label != application?.name && item.label != item.name
             glide.load(item)
                 .apply(RequestOptions().fitCenter())
                 .transition(DrawableTransitionOptions.withCrossFade())