+++ app/src/main/java/com/sdex/activityrunner/app/ApplicationsListAdapter.kt
@@ -8,14 +8,17 @@ import androidx.fragment.app.FragmentActivity
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestManager
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.bumptech.glide.request.RequestOptions
 import com.sdex.activityrunner.R
 import com.sdex.activityrunner.databinding.ItemApplicationBinding
 import com.sdex.activityrunner.db.cache.ApplicationModel
-import com.sdex.activityrunner.glide.GlideApp
 import com.sdex.activityrunner.preferences.AppPreferences
+import com.sdex.activityrunner.util.ApplicationSectionNameProvider
+import com.sdex.activityrunner.util.EmptySectionNameProvider
+import com.sdex.activityrunner.util.SectionNameProvider
 import com.simplecityapps.recyclerview_fastscroll.views.FastScrollRecyclerView
 
 class ApplicationsListAdapter(
@@ -24,11 +27,16 @@ class ApplicationsListAdapter(
 ) : ListAdapter<ApplicationModel, ApplicationsListAdapter.AppViewHolder>(DIFF_CALLBACK),
     FastScrollRecyclerView.SectionedAdapter {
 
-    private val glide = GlideApp.with(activity)
+    private val glide = Glide.with(activity)
 
     private var showSystemAppIndicator: Boolean = appPreferences.isShowSystemAppIndicator
     private var showDisabledAppIndicator: Boolean = appPreferences.isShowDisabledAppIndicator
-
+    private var sectionNameProvider: SectionNameProvider =
+        if (appPreferences.sortBy == ApplicationModel.NAME) {
+            ApplicationSectionNameProvider
+        } else {
+            EmptySectionNameProvider
+        }
     var itemClickListener: ItemClickListener? = null
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AppViewHolder {
@@ -46,13 +54,16 @@ class ApplicationsListAdapter(
         )
     }
 
-    override fun getSectionName(position: Int): String {
-        val name = getItem(position).name
-        return if (name.isNullOrEmpty()) "" else name.first().uppercaseChar().toString()
-    }
+    override fun getSectionName(position: Int): String =
+        sectionNameProvider.getSectionName(getItem(position))
 
     @SuppressLint("NotifyDataSetChanged")
     fun update() {
+        sectionNameProvider = if (appPreferences.sortBy == ApplicationModel.NAME) {
+            ApplicationSectionNameProvider
+        } else {
+            EmptySectionNameProvider
+        }
         if (appPreferences.isShowSystemAppIndicator != showSystemAppIndicator ||
             appPreferences.isShowDisabledAppIndicator != showDisabledAppIndicator
         ) {
@@ -84,19 +95,16 @@ class ApplicationsListAdapter(
             binding.packageName.text = item.packageName
             val showSystemLabel = item.system && showSystemAppIndicator
             val showDisabledLabel = !item.enabled && showDisabledAppIndicator
-            val info = StringBuilder()
-            val context = binding.root.context
-            if (showDisabledLabel) {
-                info.append(context.getString(R.string.application_disabled))
-            }
-            if (showDisabledLabel && showSystemLabel) {
-                info.append(" | ")
-            }
-            if (showSystemLabel) {
-                info.append(context.getString(R.string.application_system))
-            }
-            binding.info.text = info.toString()
             binding.info.isVisible = showDisabledLabel || showSystemLabel
+            binding.system.isVisible = showSystemLabel
+            binding.disabled.isVisible = showDisabledLabel
+
+            val context = binding.root.context
+            binding.version.text = context.getString(
+                R.string.app_version_format,
+                item.versionName,
+                item.versionCode
+            )
             glide.load(item)
                 .apply(RequestOptions().fitCenter())
                 .transition(DrawableTransitionOptions.withCrossFade())