+++ app/src/main/java/com/sdex/activityrunner/app/dialog/ApplicationOptionsDialog.kt
@@ -4,15 +4,17 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.view.ContextThemeWrapper
 import androidx.core.os.bundleOf
 import androidx.core.view.isVisible
+import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.drawable.DrawableTransitionOptions
 import com.bumptech.glide.request.RequestOptions
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import com.sdex.activityrunner.R
 import com.sdex.activityrunner.databinding.DialogApplicationMenuBinding
 import com.sdex.activityrunner.db.cache.ApplicationModel
 import com.sdex.activityrunner.extensions.serializable
-import com.sdex.activityrunner.glide.GlideApp
 import com.sdex.activityrunner.manifest.ManifestViewerActivity
 import com.sdex.activityrunner.util.AppUtils
 import com.sdex.activityrunner.util.IntentUtils
@@ -24,9 +26,14 @@ class ApplicationOptionsDialog : BottomSheetDialogFragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
+        savedInstanceState: Bundle?,
     ): View {
-        _binding = DialogApplicationMenuBinding.inflate(inflater, container, false)
+        val contextThemeWrapper = ContextThemeWrapper(activity, R.style.AppTheme)
+        _binding = DialogApplicationMenuBinding.inflate(
+            inflater.cloneInContext(contextThemeWrapper),
+            container,
+            false
+        )
         return binding.root
     }
 
@@ -35,7 +42,7 @@ class ApplicationOptionsDialog : BottomSheetDialogFragment() {
         val model = requireArguments().serializable<ApplicationModel>(ARG_MODEL)!!
         val packageName = model.packageName
 
-        GlideApp.with(this)
+        Glide.with(this)
             .load(model)
             .apply(RequestOptions().fitCenter())
             .transition(DrawableTransitionOptions.withCrossFade())