+++ app/src/main/java/com/sdex/activityrunner/util/IntentUtils.kt
@@ -14,6 +14,7 @@ import androidx.browser.customtabs.CustomTabsIntent
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.IconCompat
 import androidx.core.graphics.drawable.toBitmap
+import com.bumptech.glide.Glide
 import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.engine.GlideException
 import com.bumptech.glide.request.RequestListener
@@ -21,7 +22,6 @@ import com.bumptech.glide.request.target.Target
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.sdex.activityrunner.R
 import com.sdex.activityrunner.app.ActivityModel
-import com.sdex.activityrunner.glide.GlideApp
 import com.sdex.activityrunner.shortcut.ShortcutHandlerActivity
 import com.sdex.activityrunner.shortcut.createShortcut
 import timber.log.Timber
@@ -36,9 +36,16 @@ object IntentUtils {
         }
     }
 
-    fun createLauncherIcon(context: Context, activityModel: ActivityModel, bitmap: Bitmap?) {
+    fun createLauncherIcon(
+        context: Context,
+        activityModel: ActivityModel,
+        bitmap: Bitmap?,
+        useRoot: Boolean = false,
+    ) {
         if (bitmap != null) {
-            val intent = activityModel.toIntent(context)
+            val intent = activityModel.toIntent(context).apply {
+                putExtra(ShortcutHandlerActivity.ARG_USE_ROOT, useRoot)
+            }
             val iconCompat = try {
                 IconCompat.createWithBitmap(bitmap)
             } catch (e: Exception) { // android.os.TransactionTooLargeException
@@ -60,32 +67,35 @@ object IntentUtils {
         val intent = getActivityIntent(Intent.ACTION_VIEW, component)
         intent.putExtra(ShortcutHandlerActivity.ARG_PACKAGE_NAME, this.packageName)
         intent.putExtra(ShortcutHandlerActivity.ARG_CLASS_NAME, this.className)
-        intent.putExtra(ShortcutHandlerActivity.ARG_EXPORTED, this.exported)
         return intent
     }
 
     private fun loadActivityIcon(context: Context, activityModel: ActivityModel) {
         val activityManager = context.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager
         val launcherLargeIconSize = activityManager.launcherLargeIconSize
-        GlideApp.with(context)
+        Glide.with(context)
             .asDrawable()
             .load(activityModel)
             .error(R.mipmap.ic_launcher)
             .override(launcherLargeIconSize)
             .listener(object : RequestListener<Drawable> {
                 override fun onLoadFailed(
-                    e: GlideException?, model: Any?,
-                    target: Target<Drawable>?, isFirstResource: Boolean
+                    e: GlideException?,
+                    model: Any?,
+                    target: Target<Drawable>,
+                    isFirstResource: Boolean
                 ): Boolean {
                     return false
                 }
 
                 override fun onResourceReady(
-                    resource: Drawable?, model: Any?,
-                    target: Target<Drawable>?, dataSource: DataSource?,
+                    resource: Drawable,
+                    model: Any,
+                    target: Target<Drawable>?,
+                    dataSource: DataSource,
                     isFirstResource: Boolean
                 ): Boolean {
-                    createLauncherIcon(context, activityModel, resource?.toBitmap())
+                    createLauncherIcon(context, activityModel, resource.toBitmap())
                     return false
                 }
             })